+++ Backend/src/main/java/com/Bus_Reservation_Nov24/Bus/Reservation/Backend/Dev/Nov24/Repository/DebitCardRepo.java
@@ -3,6 +3,6 @@ package com.Bus_Reservation_Nov24.Bus.Reservation.Backend.Dev.Nov24.Repository;
 import com.Bus_Reservation_Nov24.Bus.Reservation.Backend.Dev.Nov24.Model.DebitCardDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DebitCardRepo extends JpaRepository<DebitCardDetails,Integer> {
+public interface DebitCardRepo extends JpaRepository<DebitCardDetails,Long> {
 
 }
